fix: Fix float numbers for search on import
This change detects if document contains array of objects in the schema,
and if it is the case replaces all the numbers to fractional value,
in order to guarantee that search infer their schema as float.

It inserts this record with replaced numbers for search to persists
proper schema and deletes the record immediately.

It proceeds import process as usual after that