feat: Allow to skip local TLS when token is set
By default we enforce TLS if token or secret is set.
This feature allow to disable TLS when connecting to local
instance with authenticatiaon enabled.
TLS can only be disabled for HTTP.
GRPC always enforce TLS when credentials non empty, it only
allows to disble it on Unix sockets.

* Added unix socket support.
* Fixed a bug when user was automatically logged in to local instance
  with authentication enabled.
* This also updated the client to the latest, which requires inference
code updates.